perf(web): split bundle into cached vendor/online chunks
Three improvements to bundle health and CI guardrails:

1. Code-split the online stack. Extract AppShell + the online game screen
   into their own modules and lazy()-load OnlineGameScreen, so the host
   runtime, online hook, online board and lobby dialogs leave the initial
   bundle for the common local-AI path. Add manualChunks to separate
   third-party code (vendor) from React and Sentry. A routine app deploy
   now re-downloads ~30 kB gzip (the entry chunk) instead of the previous
   single ~342 kB monolith; the rest stays cached and SW-precached.

2. Add vitest coverage thresholds in all three packages as regression
   floors a few points below current coverage. They ratchet against silent
   erosion without breaking CI on normal fluctuation.

3. Drop the two react-hooks/exhaustive-deps disables in AnimatedCard.
   Effect 1 uses honest deps (initialDelay is fixed per mount); Effect 2
   reads its inputs through a ref synced in a layout effect, so isVisible
   stays the only trigger and markAnimationStarted swapping in a new
   animation object no longer risks restarting WAAPI mid-flight.

Verified: typecheck, lint, 226 unit tests (under new thresholds), build
(chunks split as intended), and 24 local-path e2e specs all pass.
